CIF-1633: export authenticated ApolloLink
