+++ react-components/src/index.js
@@ -28,3 +28,5 @@ export { CheckoutProvider } from './components/Checkout/checkoutContext';
 
 export { CartProvider, CartInitializer } from './components/Minicart';
 export { AccountDetails } from './components/AccountDetails';
+
+export { graphqlAuthLink } from './utils/authUtils';
